refactor: rename files to keep existing ones around
Ref eng-268
