+++ packages/core/src/external/opensearch/lexical/lexical-searcher.ts
